+++ website_support/controllers/main.py
@@ -399,16 +399,25 @@ class SupportTicketController(http.Controller):
     def support_ticket_view_list(self, **kw):
         """Displays a list of support tickets owned by the logged in user"""
 
+        values = {}
+        for field_name, field_value in kw.items():
+            values[field_name] = field_value
+            
         extra_access = []
         for extra_permission in http.request.env.user.partner_id.stp_ids:
             extra_access.append(extra_permission.id)
 
-        support_tickets = http.request.env['website.support.ticket'].sudo().search(['|', ('partner_id','=',http.request.env.user.partner_id.id), ('partner_id', 'in', extra_access), ('partner_id','!=',False) ])
+        if 'state' in values:
+            support_tickets = http.request.env['website.support.ticket'].sudo().search(['|', ('partner_id','=',http.request.env.user.partner_id.id), ('partner_id', 'in', extra_access), ('partner_id','!=',False), ('state', '=', int(values['state'])) ])
+        else:
+            support_tickets = http.request.env['website.support.ticket'].sudo().search(['|', ('partner_id','=',http.request.env.user.partner_id.id), ('partner_id', 'in', extra_access), ('partner_id','!=',False) ])
 
         no_approval_required = request.env['ir.model.data'].get_object('website_support','no_approval_required')
         change_requests = http.request.env['website.support.ticket'].sudo().search(['|', ('partner_id','=',http.request.env.user.partner_id.id), ('partner_id', 'in', extra_access), ('partner_id','!=',False), ('approval_id','!=',no_approval_required.id) ], order="planned_time desc")
 
-        return http.request.render('website_support.support_ticket_view_list', {'support_tickets':support_tickets,'ticket_count':len(support_tickets), 'change_requests': change_requests, 'request_count': len(change_requests)})
+        ticket_states = http.request.env['website.support.ticket.states'].sudo().search([])
+
+        return http.request.render('website_support.support_ticket_view_list', {'support_tickets':support_tickets,'ticket_count':len(support_tickets), 'change_requests': change_requests, 'request_count': len(change_requests), 'ticket_states': ticket_states})
 
     @http.route('/support/ticket/view/<ticket>', type="http", auth="user", website=True)
     def support_ticket_view(self, ticket):