+++ website_support_analytic_timesheets/models/reports.py
@@ -7,23 +7,28 @@ _logger = logging.getLogger(__name__)
 
 class ReportWebsetAupportAnalyticTimesheetsSupportTechReport(models.Model):
 
-    _name = "report.webset_support_analytic_timesheets.support_tech_report"
+    _name = "report.website_support_analytic_timesheets.strt"
 
     @api.multi
-    def render_html(self, data):
-    
-        _logger.error("Render HTML Support Tech")
-        
-        model = self.env.context.get('active_model')
-        docs = self.env[model].browse(self.env.context.get('active_id'))
-        
-        docargs = {
-            'doc_ids': self.ids,
-            'doc_model': model,
-            'data': data['form'],
-            'docs': docs,
-            'time': time,
-            'test': 'hello',
+    def get_report_values(self, docids, data=None):
+        docs = self.env['account.analytic.line'].browse(docids)
+
+
+        date_dict = {}
+        for timesheet_line in docs:
+            #Group by Date
+            if str(timesheet_line.date) not in date_dict:
+                date_dict[str(timesheet_line.date)] = {}
+
+            #Sub group by employee name
+            if str(timesheet_line.employee_id.name) not in date_dict[str(timesheet_line.date)]:
+                date_dict[str(timesheet_line.date)][str(timesheet_line.employee_id.name)] = []
+            
+            date_dict[str(timesheet_line.date)][str(timesheet_line.employee_id.name)].append(timesheet_line)
+
+        _logger.error(date_dict)
+
+        return {
+            'doc_model': 'account.analytic.line',
+            'docs': date_dict
         }
-        
-        return self.env['report'].render('website_support_analytics_timesheets.support_tech_report_template', docargs)